gridMaker: Generalized version for lonlat points
Now the code needs the longitude and latitude inputs as a sfc_POINTS
grid. This makes it easier to work with non-regular distributed grids,
like the ones from CORDEX. The functionality should be the same for
regular-spaced grids.

Also, to create the sfc_POINTS object from the original coordinates,
please use the "coords2sf" new function.

NOTE: The code was updated as well to improve readability