Selecting grids: Improved cropping
With the improved cropping, the correct amount of grid cell that
intersects a polygon will be retrieved. Before this couldn't be achieved
because the crop funciton would only consider the centroids of the
grid cells (since the grids are created from this). Now it is solved by
adding half of the spatial resolution to the cropped bbox area.